+++ packages/core/dialect/mssql/index.ts
@@ -101,8 +101,7 @@ export default class MSSQL extends GenericDialect<MSSQLLib.ConnectionPool> imple
               tableDatabase: obj.dbName,
               tableSchema: obj.tableSchema,
             } as DatabaseInterface.Table;
-          })
-          .sort();
+          });
       });
   }
 
@@ -116,8 +115,7 @@ export default class MSSQL extends GenericDialect<MSSQLLib.ConnectionPool> imple
             obj.size = obj.size !== null ? parseInt(obj.size, 10) : null;
             obj.tableDatabase = obj.dbName;
             return obj as DatabaseInterface.TableColumn;
-          })
-          .sort();
+          });
       });
   }
 }