+++ packages/core/dialect/pgsql/queries.ts
@@ -4,17 +4,32 @@ export default {
   describeTable: `SELECT * FROM INFORMATION_SCHEMA.COLUMNS
       WHERE table_name = ':table'
         AND TABLE_SCHEMA NOT IN ('pg_catalog', 'information_schema')`,
-  fetchColumns: `SELECT TABLE_NAME AS tableName,
-        COLUMN_NAME AS columnName,
-        DATA_TYPE AS type,
-        CHARACTER_MAXIMUM_LENGTH AS size,
-        TABLE_CATALOG AS tableCatalog,
-        TABLE_SCHEMA AS tableSchema,
-        current_database() as dbName,
-        COLUMN_DEFAULT AS defaultValue,
-        IS_NULLABLE AS isNullable
-      FROM INFORMATION_SCHEMA.COLUMNS
-      WHERE TABLE_SCHEMA NOT IN ('pg_catalog', 'information_schema')`,
+  fetchColumns: `
+SELECT
+  C.TABLE_NAME AS tableName,
+  C.COLUMN_NAME AS columnName,
+  C.DATA_TYPE AS type,
+  C.CHARACTER_MAXIMUM_LENGTH AS size,
+  C.TABLE_CATALOG AS tableCatalog,
+  C.TABLE_SCHEMA AS tableSchema,
+  current_database() as dbName,
+  C.COLUMN_DEFAULT AS defaultValue,
+  C.IS_NULLABLE AS isNullable,
+  TC.constraint_type AS keytype
+FROM
+  INFORMATION_SCHEMA.COLUMNS C
+  LEFT JOIN information_schema.key_column_usage KC ON KC.table_name = C.table_name
+  AND KC.table_schema = C.table_schema
+  AND KC.column_name = C.column_name
+  LEFT JOIN information_schema.table_constraints TC ON KC.table_name = TC.table_name
+  AND KC.table_schema = TC.table_schema
+  AND KC.constraint_name = TC.constraint_name
+WHERE
+  C.TABLE_SCHEMA NOT IN ('pg_catalog', 'information_schema')
+ORDER BY
+  C.TABLE_NAME,
+  C.ORDINAL_POSITION
+`,
   fetchRecords: 'SELECT * FROM :table LIMIT :limit',
   fetchTables: `SELECT
         C.TABLE_NAME AS tableName,