+++ packages/core/utils/query.ts
@@ -1,4 +1,4 @@
-import { DatabaseInterface } from '../interface';
+import { DatabaseInterface, Settings } from '../interface';
 import { format } from '@sqltools/plugins/formatter/utils';
 
 export function parse(query = '') {
@@ -8,13 +8,13 @@ export function parse(query = '') {
 export function generateInsert(
   table: string,
   cols: Array<{ value: string, column: DatabaseInterface.TableColumn }>,
-  indentSize?: number,
+  formatOptions?: Settings['format'],
 ): string {
   let insertQuery = `INSERT INTO ${table} (${cols.map((col) => col.value).join(', ')}) VALUES (`;
   cols.forEach((col, index) => {
     insertQuery = insertQuery.concat(`'\${${index + 1}:${col.column.type}}', `);
   });
-  return format(`${insertQuery.substr(0, Math.max(0, insertQuery.length - 2))});`, indentSize)
-  .replace(/'(\${\d+:(int|bool|num).+})'/gi, '$1')
+  return format(`${insertQuery.substr(0, Math.max(0, insertQuery.length - 2))});`, formatOptions)
+  .replace(/'(\${\d+:(int|bool|num)[\w ]+})'/gi, '$1')
   .concat('$0');
 }