+++ packages/language-server/.babelrc.js
@@ -1,16 +1,16 @@
 module.exports = {
-  "presets": [
+  presets: [
     [
-      require.resolve("@babel/preset-env"),
+      require.resolve('@babel/preset-env'),
       {
-        "targets": {
-          "node": "6"
+        targets: {
+          node: '6',
         },
-      }
-    ]
+      },
+    ],
   ],
-  "plugins": [
-    require.resolve("@babel/plugin-proposal-object-rest-spread"),
-    require.resolve("@babel/plugin-transform-spread")
-  ]
-}
+  plugins: [
+    require.resolve('@babel/plugin-proposal-object-rest-spread'),
+    require.resolve('@babel/plugin-transform-spread'),
+  ],
+};