+++ packages/plugins/formatter/utils.test.ts
@@ -66,3 +66,18 @@ where
   id = $1`
   );
 });
+
+it('Format vscode query snipper', () => {
+  expect(format("INSERT INTO  access_token (created, id, scopes, ttl, userid) VALUES  ('${1:timestamp with time zone}', '${2:text}', '${3:text}', '${4:integer}', '${5:integer}');")).toEqual(
+`INSERT INTO
+  access_token (created, id, scopes, ttl, userid)
+VALUES
+  (
+    '\${1:timestamp with time zone}',
+    '\${2:text}',
+    '\${3:text}',
+    '\${4:integer}',
+    '\${5:integer}'
+  );`
+  );
+});