+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## v0.17
 
+### v0.17.7
+
+* **NEW** 🎉
+  * Added setting `format.reservedWordCase` allowing to change SQL reserved words case to `upper` or `lower`. Default is null meaning no changes.
+  * Added settings `sortColumns` to change column sorting on explorer. Default sort is by `name`.
+  * Added Primary key ![Alt text](https://raw.githubusercontent.com/mtxr/vscode-sqltools/master/packages/extension/icons/pk.svg?sanitize=true) and foreign key ![Alt text](https://raw.githubusercontent.com/mtxr/vscode-sqltools/master/packages/extension/icons/fk.svg?sanitize=true) icons for PostgreSQL columns.
+
 ### v0.17.6
 
 * **NEW** 🎉